backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
openai
pymongo[srv]
pycryptodome
mongoengine
werkzeug